+++ components/collector/src/source_collectors/dependency_track/source_up_to_dateness.py
@@ -45,6 +45,8 @@ class DependencyTrackSourceUpToDateness(DependencyTrackBase, TimePassedCollector
                     last_bom_import="" if last_bom_import is None else last_bom_import.isoformat(),
                     project=project["name"],
                     project_landing_url=f"{landing_url}/projects/{uuid}",
+                    version=project.get("version", ""),
+                    is_latest="true" if self._is_latest(project) else "false",
                 )
                 entities.append(entity)
         return entities
@@ -58,6 +60,13 @@ class DependencyTrackSourceUpToDateness(DependencyTrackBase, TimePassedCollector
         if "last BOM analysis" in event_types and (datetime := self._last_bom_analysis_datetime(project)):
             datetimes.append(datetime)
         return datetimes
+
+    def _project_matches(self, project: DependencyTrackProject, names: list[str], versions: list[str]) -> bool:
+        """Return whether the project name matches the project names and versions."""
+        only_include_latest_project_version = self._parameter("only_include_latest_project_versions")
+        if only_include_latest_project_version == "yes" and not self._is_latest(project):
+            return False
+        return super()._project_matches(project, names, versions)
 
     @classmethod
     def _last_bom_import_datetime(cls, project: DependencyTrackProject) -> datetime | None:
@@ -73,3 +82,8 @@ class DependencyTrackSourceUpToDateness(DependencyTrackBase, TimePassedCollector
     def _timestamp_to_datetime(timestamp: int | None) -> datetime | None:
         """Convert the timestamp to a datetime."""
         return None if timestamp is None else datetime_from_timestamp(timestamp)
+
+    @staticmethod
+    def _is_latest(project: DependencyTrackProject) -> bool:
+        """Return whether the project is the latest version."""
+        return project.get("isLatest", False)